requirements.txt
grpcio
grpcio-tools
redis
numpy
